back/requirements.txt
djangorestframework
django-filter
django
django-cors-headers
